+++ Source/CDSymbolsGeneratorVisitor.m
@@ -1,5 +1,5 @@
 /********************************************
-  Copyright 2016 PreEmptive Solutions, LLC
+  Copyright 2016-2017 PreEmptive Solutions, LLC
   See LICENSE.txt for licensing information
 ********************************************/