+++ CHANGELOG.md
@@ -13,7 +13,8 @@ Change Log
 ### Changes:
 
 * New identifiers will typically be eight characters long, regardless of their original length, to help prevent conflicts.
-
+* The `--translate-dsym` functionality was migrated to PPiOS-ControlFlow (See README.md for more information).
+* Removed dependency on ahocorasick library.
 
 1.0.1
 -----------------------