+++ CHANGELOG.md
@@ -7,14 +7,14 @@ Change Log
 ### Enhancements:
 
 * Dynamic frameworks can now be processed.
+* New identifiers will typically be eight characters long, regardless of their original length, to improve obfuscation and help prevent conflicts.
 * Added the git commit hash to the version information.
 * Improved documentation.
 
 ### Changes:
 
-* New identifiers will typically be eight characters long, regardless of their original length, to help prevent conflicts.
 * The `--translate-dsym` functionality was migrated to PPiOS-ControlFlow (See README.md for more information).
-* Removed dependency on ahocorasick library.
+* Removed dependency on the ahocorasick library.
 
 1.0.1
 -----------------------
@@ -59,3 +59,6 @@ Change Log
 ### Additional Details
 
 Original commit [509591f](https://github.com/Polidea/ios-class-guard/commit/509591f78f37905913ba0cbd832e5e4f7b925a8a) was corrupted. This was fixed by modifying it and rewriting the commit history after it. The new repaired commit is 496ae586. The fork point in the new history is 94121d10.
+
+---------------------------------------------------------------------
+Copyright 2016-2017 PreEmptive Solutions, LLC